+++ Dip/DipTests/Package.swift
@@ -1,0 +1,9 @@
+import PackageDescription
+
+let package = Package(
+        name: "DipTests",
+        dependencies: [
+                .Package(url: "../../../Dip", majorVersion: 4, minor: 2),
+        ]
+)
+