+++ Package.swift
@@ -27,3 +27,4 @@ import PackageDescription
 let package = Package(
     name: "Dip"
 )
+